adjustments for mobile, some refactoring
